fixed some unimportant warnings in UG_ASSERTS which occurred on linux systems.

+++ ugbase/lib_algebra/cpu_algebra/sparsematrix.h
@@ -212,7 +212,7 @@ public:
 			static value_type v(0.0);
 			return v;
 		}
-        UG_ASSERT(cols[j]==c && j >= rowStart[r] && j < rowEnd[r], "");
+        UG_ASSERT(cols[j]==(int)c && j >= (int)rowStart[r] && j < (int)rowEnd[r], "");
         return values[j];
     }
 
@@ -227,7 +227,7 @@ public:
 	value_type &operator() (size_t r, size_t c)
 	{
 		int j=get_index(r, c);
-        UG_ASSERT(j != -1 && cols[j]==c && j >= rowStart[r] && j < rowEnd[r], "");
+        UG_ASSERT(j != -1 && cols[j]==(int)c && j >= (int)rowStart[r] && j < (int)rowEnd[r], "");
         return values[j];
     }
